refactor: Исправляет наименование файла new-trip-point-presenter, согласно критерию Б-12
