+++ src/presenter/new-trip-point-presenter.js
